pyproject.toml
[project]
name = "finance-research"
version = "0.1.0"
description = "Academic research data"
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "SQLAlchemy>=2.0.0",
    "wrds>=3.0.0",
    "python-dotenv==1.1.1"
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
